+++ src/client/EnkaClient.ts
@@ -246,11 +246,10 @@ class EnkaClient {
     }
 
     /**
-     * @param playableOnly Whether to omit test avatars and other non-playable characters
-     * @returns all character data
+     * @returns all playable character data
      */
-    getAllCharacters(playableOnly = true): CharacterData[] {
-        return this.cachedAssetsManager.getGenshinCacheData("AvatarExcelConfigData").map(c => characterUtils.getCharactersById(c.id as number, this)).map(chars => chars.filter(c => !playableOnly || (playableOnly && c.isPlayable))).reduce((a, b) => [...a, ...b]);
+    getAllCharacters(): CharacterData[] {
+        return this.cachedAssetsManager.getGenshinCacheData("AvatarExcelConfigData").filter(p => p.has("useType") && p.getAsString("useType") === "AVATAR_FORMAL").map(p => characterUtils.getCharactersById(p.getAsNumber("id"), this)).reduce((a, b) => [...a, ...b], []);
     }
 
     /**
@@ -269,9 +268,9 @@ class EnkaClient {
     getAllWeapons(excludeInvalidWeapons = true): WeaponData[] {
         const weapons = this.cachedAssetsManager.getGenshinCacheData("WeaponExcelConfigData");
         if (excludeInvalidWeapons) {
-            return weapons.filter(w => w.weaponPromoteId === w.id).map(w => new WeaponData(w.id as number, this, w));
+            return weapons.filter(p => p.has("id") && p.has("weaponPromoteId") && p.getAsNumber("weaponPromoteId") === p.getAsNumber("id")).map(p => new WeaponData(p.getAsNumber("id"), this, p));
         } else {
-            return weapons.map(w => new WeaponData(w.id as number, this, w));
+            return weapons.map(p => new WeaponData(p.getAsNumber("id"), this, p));
         }
     }
 
@@ -288,7 +287,7 @@ class EnkaClient {
      * @returns all costume data
      */
     getAllCostumes(includeDefaults = false): Costume[] {
-        return this.cachedAssetsManager.getGenshinCacheData("AvatarCostumeExcelConfigData").filter(c => !includeDefaults || (includeDefaults && c.isDefault)).map(c => new Costume(null, this, c));
+        return this.cachedAssetsManager.getGenshinCacheData("AvatarCostumeExcelConfigData").filter(p => !includeDefaults || (includeDefaults && (p.has("isDefault") ? p.getAsBoolean("isDefault") : false))).map(p => new Costume(null, this, p));
     }
 
     /**
@@ -303,7 +302,7 @@ class EnkaClient {
      * @returns all material data
      */
     getAllMaterials(): Material[] {
-        return this.cachedAssetsManager.getGenshinCacheData("MaterialExcelConfigData").map(m => Material.getMaterialById(m.id as number, this, m));
+        return this.cachedAssetsManager.getGenshinCacheData("MaterialExcelConfigData").map(p => Material.getMaterialById(p.getAsNumber("id"), this, p));
     }
 
     /**
@@ -318,7 +317,7 @@ class EnkaClient {
      * @returns all name card data
      */
     getAllNameCards(): NameCard[] {
-        return this.cachedAssetsManager.getGenshinCacheData("MaterialExcelConfigData").filter(m => m.materialType === NameCard.MATERIAL_TYPE).map(n => new NameCard(n.id as number, this, n));
+        return this.cachedAssetsManager.getGenshinCacheData("MaterialExcelConfigData").filter(p => p.has("materialType") && p.getAsString("materialType") === NameCard.MATERIAL_TYPE).map(p => new NameCard(p.getAsNumber("id"), this, p));
     }
 
     /**
@@ -334,22 +333,22 @@ class EnkaClient {
      * @returns all artifact data
      */
     getAllArtifacts(highestRarityOnly = false): ArtifactData[] {
-        const excludeSetIds = this.cachedAssetsManager.getGenshinCacheData("ReliquarySetExcelConfigData").filter(s => s.DisableFilter === 1).map(s => s.setId);
+        const excludeSetIds = this.cachedAssetsManager.getGenshinCacheData("ReliquarySetExcelConfigData").filter(p => p.getValue("DisableFilter") === 1).map(p => p.getAsNumber("setId"));
 
         // including artifacts with invalid rarity
-        const artifacts = this.cachedAssetsManager.getGenshinCacheData("ReliquaryExcelConfigData").filter(a => a.setId && !excludeSetIds.includes(a.setId));
+        const artifacts = this.cachedAssetsManager.getGenshinCacheData("ReliquaryExcelConfigData").filter(p => p.has("setId") && !excludeSetIds.includes(p.getAsNumber("setId")));
 
-        const validRarityArtifacts = artifacts.filter(a => {
-            const allowedRarityRange = artifactRarityRangeMap[a.setId as number] ?? [4, 5];
+        const validRarityArtifacts = artifacts.filter(p => {
+            const allowedRarityRange = artifactRarityRangeMap[p.getAsNumber("setId")] ?? [4, 5];
             const min = highestRarityOnly ? allowedRarityRange[1] : allowedRarityRange[0];
             const max = allowedRarityRange[1];
-            const stars = a.rankLevel as number;
+            const stars = p.getAsNumber("rankLevel");
             return (min <= stars && stars <= max);
         });
 
-        const chunked = separateByValue(validRarityArtifacts, (a) => `${a.setId}-${a.equipType}-${a.rankLevel}`);
+        const chunked = separateByValue(validRarityArtifacts, (p) => `${p.getAsNumber("setId")}-${p.getAsString("equipType")}-${p.getAsNumber("rankLevel")}`);
 
-        return Object.values(chunked).map(chunk => new ArtifactData(chunk[chunk.length - 1].id as number, this));
+        return Object.values(chunked).map(chunk => new ArtifactData(chunk[chunk.length - 1].getAsNumber("id"), this));
     }
 
     /**
@@ -364,8 +363,8 @@ class EnkaClient {
      * @returns all artifact set data
      */
     getAllArtifactSets(): ArtifactSet[] {
-        const sets = this.cachedAssetsManager.getGenshinCacheData("ReliquarySetExcelConfigData").filter(s => s.DisableFilter !== 1);
-        return sets.map(s => new ArtifactSet(s.setId as number, this, s));
+        const sets = this.cachedAssetsManager.getGenshinCacheData("ReliquarySetExcelConfigData").filter(p => p.getValue("DisableFilter") !== 1);
+        return sets.map(p => new ArtifactSet(p.getAsNumber("setId"), this, p));
     }
 
     /**