+++ src/models/artifact/ArtifactSplitSubstat.ts
@@ -15,12 +15,12 @@ class ArtifactSplitSubstat extends StatusProperty {
      * @param enka
      */
     constructor(id: number, enka: EnkaClient) {
-        const data: JsonObject | undefined = enka.cachedAssetsManager.getGenshinCacheData("ReliquaryAffixExcelConfigData").find(a => a.id === id);
-        if (!data) throw new AssetsNotFoundError("Artifact Substat", id);
+        const json = enka.cachedAssetsManager.getGenshinCacheData("ReliquaryAffixExcelConfigData").find(p => p.getAsNumber("id") === id)?.detach();
+        if (!json) throw new AssetsNotFoundError("Artifact Substat", id);
 
-        super(data.propType as FightProp, data.propValue as number, enka);
+        super(json.getAsString("propType") as FightProp, json.getAsNumber("propValue"), enka);
 
-        this._data = data;
+        this._data = json.getAsJsonObject();
     }
 }