+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 3.1.0
+**This version includes Breaking Changes**
+- Removed `playableOnly` param from EnkaClient#getAllCharacters.
+- Removed CharacterData#isPlayable.
+- Safer type assertions.
 # 3.0.5
 **This version includes small Breaking Changes**
 - Made EnkaUser#uid, EnkaUser#verificationCode, EnkaUser#verificationExpires, and EnkaUser#verificationCodeRetries nullable.